refactor(src): split github/client.rs by responsibility
